+++ Firestore/CHANGELOG.md
@@ -4,6 +4,7 @@
 - [feature] Add preview support for full-text search and geo search.
 - [feature] Added support for Pipeline expressions `nor` and `switchOn`. (#15943)
 - [feature] Added support for Pipeline expressions `ifNull` and `coalesce`. (#15989)
+- [feature] Added support for Pipeline expression `isType`.
 
 # 12.11.0
 - [removed] **Breaking change:** Removed the `then(_:else:)` Beta API in `BooleanExpression` from the Pipeline Swift API. Use `ConditionalExpression` instead. (#15926)